requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
decorator==4.4.2
exceptiongroup==1.1.2
fastapi==0.99.1
flatbuffers==23.5.26
h11==0.14.0
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
lazy_loader==0.3
llvmlite==0.40.1
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==10.0.0
platformdirs==3.8.0
pooch==1.7.0
proglog==0.1.10
protobuf==4.23.3
pydantic==1.10.10
PyMatting==1.1.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyWavelets==1.4.1
rembg==2.0.49
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tifffile==2023.4.12
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
zipp==3.15.0
